+++ activerecord/lib/active_record/transactions.rb
@@ -294,7 +294,7 @@ module ActiveRecord
             fire_on = Array(options[:on])
             assert_valid_transaction_action(fire_on)
             options[:if] = Array(options[:if])
-            options[:if] << "transaction_include_any_action?(#{fire_on})"
+            options[:if] << lambda { transaction_include_any_action?(fire_on) }
           end
         end